refactor: update chainId to uint64 in tests

+++ test/Permit2Passage.t.sol
@@ -34,11 +34,11 @@ contract PassagePermit2Test is SharedPermit2Test {
     // token consts
     address token;
     uint256 amount = 200;
-    uint256 chainId = 3;
+    uint64 chainId = 3;
     address recipient = address(0x123);
 
     event EnterToken(
-        uint256 indexed rollupChainId, address indexed rollupRecipient, address indexed token, uint256 amount
+        uint64 indexed rollupChainId, address indexed rollupRecipient, address indexed token, uint256 amount
     );
 
     function setUp() public {
@@ -57,7 +57,7 @@ contract PassagePermit2Test is SharedPermit2Test {
         _setupSinglePermit(token, amount);
 
         // deploy Passage
-        target = new Passage(block.chainid + 1, address(this), initialEnterTokens, address(permit2Contract));
+        target = new Passage(uint64(block.chainid + 1), address(this), initialEnterTokens, address(permit2Contract));
         vm.label(address(target), "passage");
 
         // construct Enter witness
